+++ src/scripts/generateAutocompleteGraph.ts
@@ -7,10 +7,7 @@ import { writeFileSync } from 'fs';
 import { DirectedGraph } from 'graphology';
 import aggregatedDataRaw from '../data/aggregated_data.json';
 import professor_to_alias from '../data/professor_to_alias.json';
-import {
-  decodeSearchQueryLabel,
-  type SearchQuery,
-} from '../modules/SearchQuery/SearchQuery';
+import { decodeSearchQueryLabel, type SearchQuery } from '../utils/SearchQuery';
 
 interface ProfessorData {
   first_name?: string;
