+++ src/utils/GenericFetchedData.ts
